+++ src/code/browser/browser.ts
@@ -1,9 +1,9 @@
 import { ComponentService, IComponentService } from "src/code/browser/service/componentService";
 import { Workbench } from "src/code/browser/workbench/workbench";
 import { FileLogService, IFileLogService } from "src/code/common/service/logService/fileLogService";
-import { ServiceDescriptor } from "src/code/common/service/instantiation/descriptor";
-import { IInstantiationService, InstantiationService } from "src/code/common/service/instantiation/instantiation";
-import { ServiceCollection } from "src/code/common/service/instantiation/serviceCollection";
+import { ServiceDescriptor } from "src/code/common/service/instantiationService/descriptor";
+import { IInstantiationService, InstantiationService } from "src/code/common/service/instantiationService/instantiation";
+import { ServiceCollection } from "src/code/common/service/instantiationService/serviceCollection";
 
 
 /**
