+++ src/code/browser/workbench/actionView/actionView.ts
@@ -2,10 +2,10 @@ import { getSvgPathByName, SvgType } from 'src/base/common/string';
 import { Component, ComponentType, IComponent } from 'src/code/browser/workbench/component';
 import { ExplorerViewComponent } from "src/code/browser/workbench/actionView/explorer/explorer";
 import { EVENT_EMITTER } from 'src/base/common/event';
-import { createDecorator } from 'src/code/common/service/instantiation/decorator';
+import { createDecorator } from 'src/code/common/service/instantiationService/decorator';
 import { IComponentService } from 'src/code/browser/service/componentService';
 import { IContextMenuService } from 'src/code/browser/service/contextMenuService';
-import { IInstantiationService } from 'src/code/common/service/instantiation/instantiation';
+import { IInstantiationService } from 'src/code/common/service/instantiationService/instantiation';
 
 export type ActionViewType = 'none' | 'explorer' | 'outline' | 'search' | 'git';
