+++ src/code/common/service/instantiationService/instantiation.ts
@@ -1,8 +1,8 @@
-import { createDecorator, ServiceIdentifier, _ServiceUtil } from "src/code/common/service/instantiation/decorator";
-import { Graph, Node } from "src/code/common/service/instantiation/dependencyGraph";
-import { ServiceDescriptor } from "src/code/common/service/instantiation/descriptor";
-import { IdleValue } from "src/code/common/service/instantiation/idle";
-import { ServiceCollection } from "src/code/common/service/instantiation/serviceCollection";
+import { createDecorator, ServiceIdentifier, _ServiceUtil } from "src/code/common/service/instantiationService/decorator";
+import { Graph, Node } from "src/code/common/service/instantiationService/dependencyGraph";
+import { ServiceDescriptor } from "src/code/common/service/instantiationService/descriptor";
+import { IdleValue } from "src/code/common/service/instantiationService/idle";
+import { ServiceCollection } from "src/code/common/service/instantiationService/serviceCollection";
 
 export const IInstantiationService = createDecorator<IInstantiationService>('instantiation-service');
