+++ src/code/common/service/logService/fileLogService.ts
@@ -1,6 +1,6 @@
 import { ILogService, LogInfo, LogPathType, LogService } from "src/code/common/service/logService/logService";
 import { INoteBookManagerService } from "src/code/common/model/notebookManager";
-import { createDecorator } from "src/code/common/service/instantiation/decorator";
+import { createDecorator } from "src/code/common/service/instantiationService/decorator";
 
 export const IFileLogService = createDecorator<IFileLogService>('file-log-service');
